Move the JustLightTheme class to a seperate class file following the namespace path, so we can use the classloader.
We need this because the JustFancyTheme class (our base theme) can not be reached because the themes are loaded in alphabetical order by webtrees. At the time the JustBlack theme is loaded our base theme class doesn't exist yet, which results in a fatal error. This doesn't effect this theme but we will use the same approach for all our themes.