Entry for .markdown p contained twice
Since there is an intensive discussion in the webtrees forum, I noticed that two entries for .markdown p are contained twice in the css file. Since you have just corrected one yourself and there is a conflict with the Vestamodul ClassicLAF, I have also deleted the other one as a test. Thanks for the great theme.